fix cursor based pagination next URI
We don't want cursor based pagination to return a next URI if there
isn't any item left to be shown.

The offset based pagination already does this by running two SQL
queries: one for retrieving the actual items and the other to count the
total number of items.

This commit adds a SQL query to the cursor based pagination to count the
number of items after the cursor. If the count is lower or equal to the
page size, there isn't any item left and the next URI is null.