fix cursor base pagination next URI
We don't want cursor based pagination to return a next URI if there
isn't any item left to be shown.

The offset based pagination already does this by running two SQL
queries: one for retrieving the actual items and the other to count the
total number of items.

This commit increases the limit of the cursor based pagination by one.
If the number of items is lower than or equal to the request limit,
there isn't any item left and the next URI is null.